Absract node service
