+++ src/Pioneer.Pagination/NextPageService.cs
@@ -3,7 +3,6 @@ using System.Linq;
 
 namespace Pioneer.Pagination
 {
-
     public interface INextPageService
     {
         NextPage BuildNextPage(
@@ -17,6 +16,13 @@ namespace Pioneer.Pagination
 
     public class NextPageService : INextPageService
     {
+        private readonly ILastPageInCollectionService _pageInCollectionService;
+
+        public NextPageService()
+        {
+            _pageInCollectionService = new LastPageInCollectionService();
+        }
+
         /// <summary>
         /// Build next page object
         /// </summary>
@@ -39,26 +45,9 @@ namespace Pioneer.Pagination
         /// <summary>
         /// Determine if we need a Next Page
         /// </summary>
-        private static bool DisplayNextPage(int collectionSize, int selectedPageNumber, int itemsPerPage)
+        private bool DisplayNextPage(int collectionSize, int selectedPageNumber, int itemsPerPage)
         {
-            return selectedPageNumber < GetLastPageInCollection(collectionSize, itemsPerPage);
-        }
-
-        /// <summary>
-        /// Get last page number in collection
-        /// </summary>
-        private static int GetLastPageInCollection(int collectionSize, int itemsPerPage)
-        {
-            var lastPage = (double)collectionSize / itemsPerPage;
-
-            // If whole number, we know nothing exist past lastPage
-            if (lastPage % 1 == 0)
-            {
-                return (int)lastPage;
-            }
-
-            // If not whole number, we know there should be one more page past lastPage
-            return (int)lastPage + 1;
+            return selectedPageNumber < _pageInCollectionService.GetLastPageInCollection(collectionSize, itemsPerPage);
         }
     }
 }
